Changed cmd_await_jobs and create_job_header_string functions so that jobID is a compulsory second argument rather than an option.  This was done to avoid confusing situations where the #BSUB -w option would contain IDs generated from a hash of only part of the summary file. Created the get_groupparents function for retrieveing parent jobs of a group from a dictionary of job group arrays by looking for the group tag (default #JGROUP). The cmd_await_jobs now works by calling get_groupparents which now correctly processes the jobDate option, added unit tests for this. Added checks for non-unique group names (in split_summary) and for a group being labeled as it's own parent (in get_groupparents).
To do: The function create_jobs_from_summary_ should return an array of
dictionaries with job groups linked to output file paths.  This can
subsequently be ordered using get_prioritiesarray before being written
as an ordered list to a file.

To do: Finish the get_prioritiesarray function. Ensure that jobs which
are required by subsequent jobs come first in the list (fix the
incorrect -w option in the *_last job of the jgroup.sh unit test).